refactor: reorganize encounter setup components
